+++ EnhanceQoL/Locales/deDE.lua
@@ -54,7 +54,7 @@ L["Misc"] = "Sonstiges"
 L["Quest"] = "Quest"
 
 L["hideBagsBar"] = "Taschenleiste ausblenden"
-
+L["hideMicroMenu"] = "Mikromenü ausblenden"
 -- Dungeon
 L["autoChooseDelvePower"] = "Delve-Power automatisch auswählen, wenn nur 1 Option verfügbar ist"
 L["lfgSortByRio"] = "Sortiere Bewerber für Mythisch+ Dungeons nach Mythisch-Wertung"