+++ EnhanceQoL/Locales/ptBR.lua
@@ -55,7 +55,7 @@ L["Misc"] = "Diversos"
 L["Quest"] = "Missão"
 
 L["hideBagsBar"] = "Ocultar barra de sacolas"
-
+L["hideMicroMenu"] = "Ocultar micro menu"
 -- Dungeon
 L["autoChooseDelvePower"] = "Selecionar automaticamente o poder de incursão quando houver apenas 1 opção"
 L["lfgSortByRio"] = "Classificar candidatos de masmorras míticas pelo placar mítico"