+++ CHANGELOG.md
@@ -1,2 +1,2 @@
-## Bugfix
-- Rare error when opening dungeon frame that a toy was not found
+## Added
+- Option to Hide the Micro Menu